Add BulgeSegment class to jts-lab
The bulge is the tangent of 1/4 of the included angle for the arc
between the selected vertex and the next vertex in the polyline's
vertex list.
A negative bulge value indicates that the arc goes clockwise from the
selected vertex to the next vertex.
A bulge of 0 indicates a straight segment, and a bulge of 1 is a
semicircle.

* Add Triangle.contains(Coordinate) predicate
* Add BulgeSegmentTest class

Signed-off-by: Felix Obermaier <felix.obermaier@netcologne.de>